[Ignore] Remove constant ansible doc variables to fix ansible-doc errors
